Moved train() to src directory
